review comments & gofmt
