+++ pkg/provisioner/controller.go
@@ -7,7 +7,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/kubernetes"
@@ -286,7 +285,6 @@ func (c *Controller) handleProvisionClaim(key string, obc *v1alpha1.ObjectBucket
 	ob.Spec.ClaimRef, err = claimRefForKey(key, c.libClientset)
 	ob.Spec.ReclaimPolicy = options.ReclaimPolicy
 	ob.SetFinalizers([]string{finalizer})
-	ob.SetOwnerReferences([]metav1.OwnerReference{makeOwnerReference(obc)})
 
 	obc, err = updateObjectBucketClaimPhase(c.libClientset, obc, v1alpha1.ObjectBucketClaimStatusPhaseBound, defaultRetryBaseInterval, defaultRetryTimeout)
 	if err != nil {
@@ -324,34 +322,30 @@ func (c *Controller) handleProvisionClaim(key string, obc *v1alpha1.ObjectBucket
 }
 
 // Delete or Revoke access to bucket defined by passed-in key.
+// TODO each delete should retry a few times to mitigate intermittent errors
 func (c *Controller) handleDeleteClaim(key string) error {
 	// Call `Delete` for new (greenfield) buckets with reclaimPolicy == "Delete".
 	// Call `Revoke` for new buckets with reclaimPolicy != "Delete".
 	// Call `Revoke` for existing (brownfield) buckets regardless of reclaimPolicy.
 
-	// TODO each delete should retry a few times to mitigate intermittent errors
-	// (Jon, is this TODO still needed??)
-
 	ob, cm, secret, err := c.getResourcesFromKey(key)
 	if err != nil {
 		return err
 	}
-	// Note: ob or cm or secret (or 2 of the 3) can be nil.
-	//   This allows us to delete remaining non-nil resources.
 
+	// ob or cm or secret (or 2 of the 3) can be nil, but Delete/Revoke cannot be called
+	// if the ob is nil. However if the secret and/or cm != nil we can delete them.
 	if ob == nil {
 		log.Error(nil, "nil ObjectBucket, assuming it has been deleted")
-		// can't call Delete or Revoke but try to delete secret and cm if they exist
 		return c.deleteResources(ob, cm, secret)
 	}
-
 	if ob.Spec.ReclaimPolicy == nil {
-		log.Error(nil, "null reclaimPolicy", "ob", ob.Name)
+		log.Error(nil, "missing reclaimPolicy", "ob", ob.Name)
 		return nil
 	}
 
-	// ok to call Delete or Revoke and attempt to delete generated k8s resources.
-	// Note: if Delete or Revoke return an error we do not attempt to delete resources.
+	// ok to call Delete or Revoke and then delete generated k8s resources
+	// Note: if Delete or Revoke return an error we do not try to delete resources
 	ob, err = updateObjectBucketPhase(c.libClientset, ob, v1alpha1.ObjectBucketClaimStatusPhaseReleased, defaultRetryBaseInterval, defaultRetryTimeout)
 	if err != nil {
 		return err
@@ -369,33 +363,11 @@ func (c *Controller) handleDeleteClaim(key string) error {
 		}
 	}
 
-	if err = deleteObjectBucket(ob, c.libClientset); err != nil {
-		if errors.IsNotFound(err) {
-			log.Error(err, "ObjectBucket vanished during deprovisioning, assuming deletion complete")
-		} else {
-			return fmt.Errorf("error deleting objectBucket %v", ob.Name)
-		}
-	}
-
-	// delete all generated kubernetes resources
 	return c.deleteResources(ob, cm, secret)
 }
 
 func (c *Controller) supportedProvisioner(provisioner string) bool {
 	return provisioner == c.provisionerName
-}
-
-func (c *Controller) objectBucketForClaimKey(key string) (*v1alpha1.ObjectBucket, error) {
-	logD.Info("getting objectBucket for key", "key", key)
-	name, err := obNameFromClaimKey(key)
-	if err != nil {
-		return nil, err
-	}
-	ob, err := c.libClientset.ObjectbucketV1alpha1().ObjectBuckets().Get(name, metav1.GetOptions{})
-	if err != nil {
-		return nil, fmt.Errorf("error listing object buckets: %v", err)
-	}
-	return ob, nil
 }
 
 // Returns the ob, configmap, and secret based on the passed-in key. Only returns non-nil
@@ -406,24 +378,16 @@ func (c *Controller) getResourcesFromKey(key string) (*v1alpha1.ObjectBucket, *c
 	if errors.IsNotFound(obErr) {
 		log.Error(obErr, "ObjectBucket not found")
 		obErr = nil
-	} else if obErr != nil {
-		ob = nil
 	}
-
 	cm, cmErr := configMapForClaimKey(key, c.clientset)
 	if errors.IsNotFound(cmErr) {
 		log.Error(cmErr, "ConfigMap not found")
 		cmErr = nil
-	} else if cmErr != nil {
-		cm = nil
 	}
-
 	s, sErr := secretForClaimKey(key, c.clientset)
 	if errors.IsNotFound(sErr) {
 		log.Error(sErr, "Secret not found")
 		sErr = nil
-	} else if sErr != nil {
-		s = nil
 	}
 
 	var err error
@@ -435,8 +399,11 @@ func (c *Controller) getResourcesFromKey(key string) (*v1alpha1.ObjectBucket, *c
 	return ob, cm, s, err
 }
 
-// Returns err if can't delete one or more of the resources. The final error, if non-nil, is
-// somewhat arbitrary.
+// Deleting the resources generated by a Provision or Grant call is triggered by the delete of
+// the OBC. Since the secret and configmap's ownerReference is the OBC they will be garbage
+// collected by k8s once their finalizers are removed. However, the OB must be explicitly
+// deleted since it is a global resource and cannot have a namespaced ownerReference.
+// Returns err if we can't delete one or more of the resources. The final error is somewhat arbitrary.
 func (c *Controller) deleteResources(ob *v1alpha1.ObjectBucket, cm *corev1.ConfigMap, s *corev1.Secret) (err error) {
 	if ob != nil {
 		if delErr := deleteObjectBucket(ob, c.libClientset); delErr != nil {
